requirements.txt
spotipy==2.24.0
google-cloud-aiplatform==1.65.0
google-cloud-firestore==2.18.0
openai==1.44.0
pyyaml==6.0.2
python-dotenv==1.0.1
Flask==3.0.3
Flask-Session==0.8.0
gunicorn==22.0.0
Werkzeug==3.0.3
PyYAML==6.0.2
cryptography==43.0.1
